refactor(examples): splitting src of StylingAndTheming example; split
